requirements.txt
astroid==2.4.2
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.2
mccabe==0.6.1
peewee==3.13.3
pylint==2.5.3
PyMySQL==0.10.0
python-dateutil==2.8.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
urllib3==1.25.9
wrapt==1.12.1
